+++ R/RcppExports.R
@@ -11,14 +11,15 @@
 #' @param D The discontinuation date relative to randomization.
 #' @param model_ki The model for the number of skipped
 #'   visits between two consecutive drug dispensing visits.
-#'   Options include "constant", "poisson", "zip" for zero-inflated
-#'   Poisson, and "nb" for negative binomial.
+#'   Options include "constant", "poisson", "zero-inflated poisson",
+#'   and "negative binomial".
 #' @param theta_ki The model parameters for the number of skipped visits
 #'   between two consecutive drug dispensing visits.
-#' @param muT The regression coefficient for the linear model for the
-#'   gap time between two consecutive drug dispensing visits.
-#' @param sigmaT The residual standard deviation for the linear model
-#'   for the gap time between two consecutive drug dispensing visits.
+#' @param model_ti The model for the gap time between two consecutive
+#'   drug dispensing visits. Options include "least squares"
+#'   and "least absolute deviations".
+#' @param theta_ti The model parameters for the gap time between
+#'   two consecutive drug dispensing visits.
 #'
 #' @return A data frame with two variables:
 #'
@@ -33,12 +34,12 @@
 #'
 #' f_dose_ongoing_cpp(
 #'   usubjid = "A001", V = 297, C = 329, D = 569,
-#'   model_ki = "zip", theta_ki = c(0.4, 2.5),
-#'   muT = 21, sigmaT = 2.3)
+#'   model_ki = "zero-inflated poisson", theta_ki = c(0.4, 2.5),
+#'   model_ti = "least squares", theta_ti = c(21, 2.3))
 #'
 #' @export
-f_dose_ongoing_cpp <- function(usubjid, V, C, D, model_ki, theta_ki, muT, sigmaT) {
-    .Call(`_drugDemand_f_dose_ongoing_cpp`, usubjid, V, C, D, model_ki, theta_ki, muT, sigmaT)
+f_dose_ongoing_cpp <- function(usubjid, V, C, D, model_ki, theta_ki, model_ti, theta_ti) {
+    .Call(`_drugDemand_f_dose_ongoing_cpp`, usubjid, V, C, D, model_ki, theta_ki, model_ti, theta_ti)
 }
 
 #' @title Dosing Date Imputation for New Patients
@@ -51,8 +52,8 @@ f_dose_ongoing_cpp <- function(usubjid, V, C, D, model_ki, theta_ki, muT, sigmaT
 #' @param D The discontinuation date relative to randomization.
 #' @param model_k0 The model for the number of skipped
 #'   visits between randomization and the first drug dispensing visit.
-#'   Options include "constant", "poisson", "zip" for zero-inflated
-#'   Poisson, and "nb" for negative binomial.
+#'   Options include "constant", "poisson", "zero-inflated poisson",
+#'   and "negative binomial".
 #' @param theta_k0 The model parameters for the number of skipped
 #'   visits between randomization and the first drug dispensing visit.
 #' @param model_t0 The model for the gap time between randomization
@@ -62,22 +63,23 @@ f_dose_ongoing_cpp <- function(usubjid, V, C, D, model_ki, theta_ki, muT, sigmaT
 #' @param theta_t0 The model parameters for the gap time between
 #'   randomization and the first drug dispensing visit when there is
 #'   no visit skipping.
-#' @param mu0 The regression coefficient for the linear model for the gap
-#'   time between randomization and the first drug dispensing visit
-#'   when there is visit skipping.
-#' @param sigma0 The residual standard deviation for the linear model
-#'   for the gap time between randomization and the first drug
-#'   dispensing visit when there is visit skipping.
+#' @param model_t1 The model for the gap time between randomization
+#'   and the first drug dispensing visit when there is visit skipping.
+#'   Options include "least squares", and "least absolute deviations".
+#' @param theta_t1 The model parameters for the gap time between
+#'   randomization and the first drug dispensing visit when there is
+#'   visit skipping.
 #' @param model_ki The model for the number of skipped
 #'   visits between two consecutive drug dispensing visits.
-#'   Options include "constant", "poisson", "zip" for zero-inflated
-#'   Poisson, and "nb" for negative binomial.
+#'   Options include "constant", "poisson", "zero-inflated poisson",
+#'   and "negative binomial".
 #' @param theta_ki The model parameters for the number of skipped
 #'   visits between two consecutive drug dispensing visits.
-#' @param muT The regression coefficient for the linear model for the gap
-#'   time between two consecutive drug dispensing visits.
-#' @param sigmaT The residual standard deviation the linear model for the
-#'   gap time between two consecutive drug dispensing visits.
+#' @param model_ti The model for the gap time between two consecutive
+#'   drug dispensing visits. Options include "least squares"
+#'   and "least absolute deviations".
+#' @param theta_ti The model parameters for the gap time between
+#'   two consecutive drug dispensing visits.
 #'
 #' @return A data frame with two variables:
 #'
@@ -92,15 +94,15 @@ f_dose_ongoing_cpp <- function(usubjid, V, C, D, model_ki, theta_ki, muT, sigmaT
 #'
 #' f_dose_new_cpp(
 #'   usubjid = "Z001", V = 0, C = 87, D = 985,
-#'   model_k0 = "zip", theta_k0 = c(0.6, 1.1),
+#'   model_k0 = "zero-inflated poisson", theta_k0 = c(0.6, 1.1),
 #'   model_t0 = "log-logistic", theta_t0 = c(-1.0, 0.7),
-#'   mu0 = 21.5, sigma0 = 1.9,
-#'   model_ki = "zip", theta_ki = c(0.1, 0.4),
-#'   muT = 21, sigmaT = 2.3)
+#'   model_t1 = "least squares", theta_t1 = c(21.5, 1.9),
+#'   model_ki = "zero-inflated poisson", theta_ki = c(0.1, 0.4),
+#'   model_ti = "least squares", theta_ti = c(21, 2.3))
 #'
 #' @export
-f_dose_new_cpp <- function(usubjid, V, C, D, model_k0, theta_k0, model_t0, theta_t0, mu0, sigma0, model_ki, theta_ki, muT, sigmaT) {
-    .Call(`_drugDemand_f_dose_new_cpp`, usubjid, V, C, D, model_k0, theta_k0, model_t0, theta_t0, mu0, sigma0, model_ki, theta_ki, muT, sigmaT)
+f_dose_new_cpp <- function(usubjid, V, C, D, model_k0, theta_k0, model_t0, theta_t0, model_t1, theta_t1, model_ki, theta_ki, model_ti, theta_ti) {
+    .Call(`_drugDemand_f_dose_new_cpp`, usubjid, V, C, D, model_k0, theta_k0, model_t0, theta_t0, model_t1, theta_t1, model_ki, theta_ki, model_ti, theta_ti)
 }
 
 norm_rej <- function(a, b) {